requirements.txt
fastapi==0.114.0
pymongo==4.8.0
pydantic==2.9.0
discord.py==2.4.0
